requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pydantic==2.9.2
pydantic-settings==2.5.2
structlog==24.4.0

sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0          
alembic==1.13.2

opensearch-py==2.7.1

redis==5.0.8
dramatiq[redis,watch]==1.16.0
APScheduler==3.10.4

httpx==0.27.2
tenacity==9.0.0

slowapi==0.1.9
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
pyjwt==2.9.0
python-multipart==0.0.12

prometheus-client==0.21.0

opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-asyncpg==0.48b0
opentelemetry-instrumentation-redis==0.48b0
